+++ jablib/src/main/java/org/jabref/logic/importer/fetcher/GenericUrlBasedFetcher.java
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 package org.jabref.logic.importer.fetcher;
 
 import java.util.List;
@@ -6,10 +7,20 @@ import java.util.Locale;
 import org.jabref.logic.importer.FetcherException;
 import org.jabref.logic.util.URLUtil;
 import org.jabref.logic.util.strings.StringUtil;
+=======
+
+package org.jabref.logic.importer.fetcher;
+
+import java.util.List;
+
+import org.jabref.logic.importer.FetcherException;
+import org.jabref.logic.importer.UrlBasedFetcher;
+>>>>>>> origin/main
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.field.StandardField;
 import org.jabref.model.entry.types.StandardEntryType;
 
+<<<<<<< HEAD
 public class GenericUrlBasedFetcher {
 
     public List<BibEntry> fetchEntryFromUrl(String url) throws FetcherException {
@@ -36,4 +47,24 @@ public class GenericUrlBasedFetcher {
     public String getName() {
         return "Generic URL Fetcher";
     }
+=======
+public class GenericUrlBasedFetcher implements UrlBasedFetcher {
+
+    @Override
+    public String getName() {
+        return "URL";
+    }
+
+    @Override
+    public List<BibEntry> performSearch(String url) throws FetcherException {
+        if (url.isBlank()) {
+            return List.of();
+        }
+
+        String trimmedUrl = url.trim();
+
+        BibEntry entry = new BibEntry(StandardEntryType.Misc).withField(StandardField.URL, trimmedUrl);
+        return List.of(entry);
+    }
+>>>>>>> origin/main
 }